add workaround for llmc issue and add unit tests for model_free_utils helpers
This commit introduces a comprehensive suite of unit tests for the model_free_utils module, covering various helper functions including _add_routed_experts_if_moe, _build_mxfp_autoround_quantization_config, and others. The tests validate the correct behavior of these functions in different scenarios, ensuring robustness and reliability in the quantization process for model compression. Additionally, the tests include checks for the _PatternMatcher and quantization configuration building functions, contributing to improved code quality and maintainability.
